refactor: move SentenceCase to :anki-common
For the widget module extraction - used in CardAnalysisWidgetConfig

Part of 20737
